refactor: move app to `/src` directory
the root directory was starting to get cluttered with config files
