+++ .prettierignore
@@ -1,0 +1,6 @@
+.next/
+build/
+.netlify/
+out/
+node_modules/
+public/
