+++ src/components/AnimatedIntro.tsx
