+++ src/components/EduProgress/index.tsx
