+++ src/pages/colors/[slug].tsx
