+++ src/pages/colors/index.tsx
